Cleans and Refactors
