+++ AngularClient/src/app/_services/index.ts
@@ -3,3 +3,5 @@ export * from './authentication.service';
 export * from './user.service';
 export * from './instruments.service';
 export * from './uniforms.service';
+export * from './ensemble.service';
+export * from './student.service';
