+++ AngularClient/src/app/_services/student.service.ts
@@ -2,7 +2,7 @@ import { environment } from 'src/environments/environment';
 import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Student } from './_models';
+import { Student } from '../_models';
 
 @Injectable({
   providedIn: 'root'
